Add icon-art-oem-test (#40)
* Add config.yaml

* Replace inidate with startdate_sim

* Fix icontools scripts

* Fix chem file names

* Set correct input file names

* Small fix for latbc grid file name

* Add prefix to chem file

* GitHub Action: Apply Pep8-formatting

* Fix for chem file

* GitHub Action: Apply Pep8-formatting

* More corrections for chem file

* GitHub Action: Apply Pep8-formatting

* Fix art namelist

* Fix art input folder

* Update on how to build icon(-art)

* Update .gitignore

* Correcting case name in comment

* Use config variables for num_procs in namelist

* Fix ICON namelists

* Remove config.py

* Only process initial file if present

* GitHub Action: Apply Pep8-formatting

* Add start of icontools runscript to logger

---------

+++ jobs/prepare_data.py
@@ -318,7 +318,7 @@ def main(cfg, model_cfg):
                     time.strftime(cfg.meteo['nameformat']))
                 if cfg.model == 'icon-art' or cfg.model == 'icon-art-oem':
                     chem_file = os.path.join(
-                        cfg.icon_input_icbc,
+                        cfg.icon_input_icbc, cfg.chem['prefix'] +
                         time.strftime(cfg.chem_nameformat))
                     datafile_list_chem.append(chem_file + cfg.chem['suffix'])
                 if meteo_file.endswith('00'):
@@ -349,6 +349,7 @@ def main(cfg, model_cfg):
                                         datafile_list=datafile_list,
                                         datafile_list_rest=datafile_list_rest,
                                         datafile_list_chem=datafile_list_chem))
+                logging.info(f" Starting icontools runscript {runscript}.")
                 result = subprocess.run([
                     "sbatch", "--wait",
                     os.path.join(cfg.icon_work, "%s.job" % runscript)
@@ -357,7 +358,7 @@ def main(cfg, model_cfg):
                 if exitcode != 0:
                     raise RuntimeError(
                         "sbatch returned exitcode {}".format(exitcode))
-                logging.info("%s successfully executed." % runscript)
+                logging.info(f"{runscript} successfully executed.")
 
             #-----------------------------------------------------
             # Add GEOSP to all meteo files
@@ -414,33 +415,34 @@ def main(cfg, model_cfg):
                     cfg.startdate_sim.strftime(cfg.meteo['prefix'] +
                                                cfg.meteo['nameformat']) +
                     '.nc')
-                merged_file = os.path.join(
-                    cfg.icon_input_icbc,
-                    cfg.startdate_sim.strftime(cfg.meteo['prefix'] +
-                                               cfg.meteo['nameformat']) +
-                    '_merged.nc')
-                ds = xr.open_dataset(meteo_file)
-                merging = False
-                if 'PS' not in ds:
-                    if 'LNPS' not in ds:
-                        raise KeyError(
-                            f"'LNPS' must be found in the initial conditions file {meteo_file}"
-                        )
-                    merging = True
-                    ds['PS'] = ds['LNPS']
-                    ds['PS'].attrs = ds['LNPS'].attrs
-                    ds['PS'] = np.exp(ds['PS'])
-                    ds['PS'] = ds['PS'].squeeze(dim='lev_2')
-                    ds['PS'].attrs["long_name"] = 'surface pressure'
-                    ds['PS'].attrs['units'] = 'Pa'
-                    logging.info(f"Added PS to file {meteo_file}")
-                if 'Q' not in ds:
-                    merging = True
-                    ds['Q'] = ds['QV']
-                    logging.info(f"Added Q to file {meteo_file}")
-                if merging:
-                    ds.to_netcdf(merged_file)
-                    tools.rename_file(merged_file, meteo_file)
+                if os.path.isfile(meteo_file):
+                    merged_file = os.path.join(
+                        cfg.icon_input_icbc,
+                        cfg.startdate_sim.strftime(cfg.meteo['prefix'] +
+                                                   cfg.meteo['nameformat']) +
+                        '_merged.nc')
+                    ds = xr.open_dataset(meteo_file)
+                    merging = False
+                    if 'PS' not in ds:
+                        if 'LNPS' not in ds:
+                            raise KeyError(
+                                f"'LNPS' must be found in the initial conditions file {meteo_file}"
+                            )
+                        merging = True
+                        ds['PS'] = ds['LNPS']
+                        ds['PS'].attrs = ds['LNPS'].attrs
+                        ds['PS'] = np.exp(ds['PS'])
+                        ds['PS'] = ds['PS'].squeeze(dim='lev_2')
+                        ds['PS'].attrs["long_name"] = 'surface pressure'
+                        ds['PS'].attrs['units'] = 'Pa'
+                        logging.info(f"Added PS to file {meteo_file}")
+                    if 'Q' not in ds:
+                        merging = True
+                        ds['Q'] = ds['QV']
+                        logging.info(f"Added Q to file {meteo_file}")
+                    if merging:
+                        ds.to_netcdf(merged_file)
+                        tools.rename_file(merged_file, meteo_file)
 
             #-----------------------------------------------------
             # In case of OEM: merge chem tracers with meteo-files
@@ -457,31 +459,34 @@ def main(cfg, model_cfg):
                             cfg.icon_input_icbc,
                             time.strftime(cfg.meteo['prefix'] +
                                           cfg.meteo['nameformat']) + '.nc')
-                        chem_file = os.path.join(
-                            cfg.icon_input_icbc,
-                            time.strftime(cfg.chem_nameformat) + '.nc')
-                        merged_file = os.path.join(
-                            cfg.icon_input_icbc,
-                            time.strftime(cfg.meteo['prefix'] +
-                                          cfg.meteo['nameformat']) +
-                            '_merged.nc')
-                        ds_meteo = xr.open_dataset(meteo_file)
-                        ds_chem = xr.open_dataset(chem_file)
-                        # LNPS --> PS
-                        ds_chem['PS'] = ds_chem['LNPS']
-                        ds_chem['PS'].attrs = ds_chem['LNPS'].attrs
-                        ds_chem['PS'] = ds_chem['PS'].squeeze(dim='lev_2')
-                        ds_chem['PS'].attrs["long_name"] = 'surface pressure'
-                        # merge:
-                        ds_merged = xr.merge([ds_meteo, ds_chem],
-                                             compat="override")
-                        #ds_merged.attrs = ds.attrs
-                        ds_merged.to_netcdf(merged_file)
-                        # Rename file to get original file name
-                        tools.rename_file(merged_file, meteo_file)
-                        tools.remove_file(chem_file)
-                        logging.info("Added chemical tracer to file {}".format(
-                            merged_file))
+                        if os.path.isfile(meteo_file):
+                            chem_file = os.path.join(
+                                cfg.icon_input_icbc, cfg.chem['prefix'] +
+                                time.strftime(cfg.chem['nameformat']) + '.nc')
+                            merged_file = os.path.join(
+                                cfg.icon_input_icbc,
+                                time.strftime(cfg.meteo['prefix'] +
+                                              cfg.meteo['nameformat']) +
+                                '_merged.nc')
+                            ds_meteo = xr.open_dataset(meteo_file)
+                            ds_chem = xr.open_dataset(chem_file)
+                            # LNPS --> PS
+                            ds_chem['PS'] = ds_chem['LNPS']
+                            ds_chem['PS'].attrs = ds_chem['LNPS'].attrs
+                            ds_chem['PS'] = ds_chem['PS'].squeeze(dim='lev_2')
+                            ds_chem['PS'].attrs[
+                                "long_name"] = 'surface pressure'
+                            # merge:
+                            ds_merged = xr.merge([ds_meteo, ds_chem],
+                                                 compat="override")
+                            #ds_merged.attrs = ds.attrs
+                            ds_merged.to_netcdf(merged_file)
+                            # Rename file to get original file name
+                            tools.rename_file(merged_file, meteo_file)
+                            tools.remove_file(chem_file)
+                            logging.info(
+                                "Added chemical tracer to file {}".format(
+                                    merged_file))
 
                     #------------
                     # Merge LBC:
@@ -491,7 +496,7 @@ def main(cfg, model_cfg):
                         time.strftime(cfg.meteo['prefix'] +
                                       cfg.meteo['nameformat']) + '_lbc.nc')
                     chem_file = os.path.join(
-                        cfg.icon_input_icbc,
+                        cfg.icon_input_icbc, cfg.chem['prefix'] +
                         time.strftime(cfg.chem_nameformat) + '_lbc.nc')
                     merged_file = os.path.join(
                         cfg.icon_input_icbc,
